Threading.Mutex 第一版。

+++ ZezeJava/ZezeJava/src/main/java/Zeze/Builtin/Threading/BMutexTryLockReadOnly.java
@@ -8,5 +8,5 @@ public interface BMutexTryLockReadOnly {
     BMutexTryLock copy();
 
     Zeze.Builtin.Threading.BLockName getLockName();
-    long getTimeoutMs();
+    int getTimeoutMs();
 }
